refactor: split archive commands file
